+++ src/main/kotlin/io/runescript/plugin/ide/folding/RsFoldingBuilder.kt
@@ -9,8 +9,8 @@ import com.intellij.openapi.util.TextRange
 import com.intellij.psi.PsiComment
 import com.intellij.psi.PsiElement
 import com.intellij.psi.util.PsiTreeUtil
-import com.intellij.refactoring.suggested.endOffset
-import com.intellij.refactoring.suggested.startOffset
+import com.intellij.psi.util.endOffset
+import com.intellij.psi.util.startOffset
 import io.runescript.plugin.lang.psi.*
 
 class RsFoldingBuilder : FoldingBuilderEx(), DumbAware {