+++ src/main/kotlin/io/runescript/plugin/lang/psi/mixin/RsConstantExpressionMixin.kt
@@ -4,7 +4,7 @@ import com.intellij.extapi.psi.ASTWrapperPsiElement
 import com.intellij.lang.ASTNode
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiReference
-import com.intellij.refactoring.suggested.startOffset
+import com.intellij.psi.util.startOffset
 import io.runescript.plugin.lang.psi.RsConstantExpression
 import io.runescript.plugin.lang.psi.RsPsiImplUtil
 import io.runescript.plugin.lang.psi.refs.RsConstantReference